💡 refactor : form전환 코드개선 및 type폴더와 icon경로 변경
