+++ src/assets/icon.ts
@@ -1,61 +1,62 @@
-import account from "./icons/account.png";
-import bookmarkFull from "./icons/bookmark-full.png";
-import bookmarkLine from "./icons/bookmark-line.png";
-import chevron from "./icons/chevron.png";
-import defaultProfile01 from "./icons/default-01.png";
-import defaultProfile02 from "./icons/default-02.png";
-import defaultProfile03 from "./icons/default-03.png";
-import dish from "./icons/dish.png";
-import editUserInfo from "./icons/edit-profile.png";
-import footerSNS01 from "./icons/footer-sns-01.png";
-import footerSNS02 from "./icons/footer-sns-02.png";
-import footerSNS03 from "./icons/footer-sns-03.png";
-import formLogo from "./icons/form-logo.png";
-import formSNS01 from "./icons/form-sns-01.png";
-import formSNS02 from "./icons/form-sns-02.png";
-import formSNS03 from "./icons/form-sns-03.png";
-import Logo from "./icons/logo.png";
-import logout from "./icons/logout.png";
-import pencil from "./icons/pencil.png";
-import search from "./icons/search.png";
-import share from "./icons/share.png";
-import top from "./icons/up-arrow.png";
+import account from './icons/account.png';
+import bookmarkFull from './icons/bookmark-full.png';
+import bookmarkLine from './icons/bookmark-line.png';
+import chevron from './icons/chevron.png';
+import defaultProfile01 from './icons/default-01.png';
+import defaultProfile02 from './icons/default-02.png';
+import defaultProfile03 from './icons/default-03.png';
+import dish from './icons/dish.png';
+import editUserInfo from './icons/edit-profile.png';
+import footerSNS01 from './icons/footer-sns-01.png';
+import footerSNS02 from './icons/footer-sns-02.png';
+import footerSNS03 from './icons/footer-sns-03.png';
+import formLogo from './icons/form-logo.png';
+import LogoGroup from './icons/LogoGroup.svg';
+import formSNS01 from './icons/form-sns-01.png';
+import formSNS02 from './icons/form-sns-02.png';
+import formSNS03 from './icons/form-sns-03.png';
+import Logo from './icons/logo.png';
+import logout from './icons/logout.png';
+import pencil from './icons/pencil.png';
+import search from './icons/search.png';
+import share from './icons/share.png';
+import top from './icons/up-arrow.png';
 
 export const HeaderIcon = {
-  Logo,
-  search,
-  bookmarkLine,
-  bookmarkFull,
-  account,
+    Logo,
+    search,
+    bookmarkLine,
+    bookmarkFull,
+    account,
 };
 export const FooterIcon = {
-  Logo,
-  footerSNS01,
-  footerSNS02,
-  footerSNS03,
+    Logo,
+    footerSNS01,
+    footerSNS02,
+    footerSNS03,
 };
 export const MainIcon = {
-  chevron,
-  top,
+    chevron,
+    top,
 };
 export const FormIcon = {
-  Logo,
-  formLogo,
-  formSNS01,
-  formSNS02,
-  formSNS03,
+    LogoGroup,
+    formLogo,
+    formSNS01,
+    formSNS02,
+    formSNS03,
 };
 export const MyDialogIcon = {
-  defaultProfile01,
-  defaultProfile02,
-  defaultProfile03,
-  editUserInfo,
-  pencil,
-  logout,
+    defaultProfile01,
+    defaultProfile02,
+    defaultProfile03,
+    editUserInfo,
+    pencil,
+    logout,
 };
 export const OneRecipeIcon = {
-  bookmarkLine,
-  bookmarkFull,
-  dish,
-  share,
+    bookmarkLine,
+    bookmarkFull,
+    dish,
+    share,
 };
