+++ src/components/button.tsx
@@ -1,9 +1,10 @@
 import { FC } from 'react';
-import { ButtonTypes } from './type/button.type';
+import { ButtonTypes } from './type/components.type';
+
 
 const ChapButton: FC<ButtonTypes> = ({ children, type, variant, size, ...props }) => {
     const variantCSS = {
-        cheese: 'bg-primary-cheese shadow-dropCheese hover:shadow-innerCheese focus:shadow-innerCheese',
+        cheese: 'bg-primary-cheese text-primary-peanut shadow-dropCheese hover:shadow-innerCheese focus:shadow-innerCheese',
         peanut: 'bg-primary-peanut shadow-dropPeanut hover:shadow-innerPeanut focus:shadow-innerPeanut text-primary-cheese',
         noResults: 'bg-primary-peanut hover:bg-primary-cheese text-primary-cheese hover:text-primary-peanut',
     };
@@ -17,7 +18,7 @@ const ChapButton: FC<ButtonTypes> = ({ children, type, variant, size, ...props }
         <button
             type={type}
             {...props}
-            className={`${variantCSS[variant]} ${sizeCSS[size]} transition duration-500 ease-in-out`}
+            className={`${variantCSS[variant]} ${sizeCSS[size]} transition duration-500 ease-in-out focus:outline-none`}
         >
             {children}
         </button>
