+++ src/components/type/components.type.ts
@@ -1,4 +1,5 @@
 import { ButtonHTMLAttributes } from "react";
+import { InputHTMLAttributes } from "react";
 
 export type ButtonTypes = {
   type: "button" | "submit";
@@ -6,3 +7,10 @@ export type ButtonTypes = {
   variant: "cheese" | "peanut" | "noResults";
   size: "circle" | "rounded" | "square";
 } & ButtonHTMLAttributes<HTMLButtonElement>;
+
+export type InputTypes = {
+  label: string;
+  type: "text" | "password";
+  error?: string;
+  access?: string;
+} & InputHTMLAttributes<HTMLInputElement>;
