--- src/components/type/input.type.ts
@@ -1,8 +1,0 @@
-import { InputHTMLAttributes } from "react";
-
-export type InputTypes = {
-  label: string;
-  type: "text" | "password";
-  error?: string;
-  access?: string;
-} & InputHTMLAttributes<HTMLInputElement>;
