+++ src/pages/sign/components/overlay/type/overlay.type.ts
@@ -1,0 +1,5 @@
+import { ReactEventHandler } from 'react';
+
+export type OverlayProps = {
+    onChangeForm: ReactEventHandler;
+};
